+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -110,6 +110,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Removed the macro ENABLE_RAYTRACING for most of the ray tracing code
 - Fixed prefab containing camera reloading in loop while selected in the Project view
 - Fixed issue causing NaN wheh the Z scale of an object is set to 0.
+- Fixed DXR shader passes attempting to render before pipeline loaded
 
 ### Changed
 - Color buffer pyramid is not allocated anymore if neither refraction nor distortion are enabled